requirements.txt
amqp==5.1.1
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
environ==1.0
Flask==2.2.2
Flask-Mail==0.9.1
Flask-Script==2.0.6
greenlet==1.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
redis